# schema/openapi/openapi.yaml
openapi: '3.0.3'
info:
  title: Backbone API
  version: '1.0.0'
  description: |
    Auto-generated by backbone-cli.
    This API provides 11 standard CRUD endpoints for each model.
servers:
  - url: http://localhost:3000
    description: Development server
...
paths:
  /api/v1/campaigns:
    get:
      tags:
        - Campaigns
      summary: List Campaigns
      description: Get a paginated list of Campaigns with optional filtering and sorting
      operationId: listCampaigns
      parameters:
        - $ref: '#/components/parameters/PageParam'
        - $ref: '#/components/parameters/PerPageParam'
        - $ref: '#/components/parameters/SortByParam'
        - $ref: '#/components/parameters/SortOrderParam'
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CampaignList'
        '401':
          $ref: '#/components/responses/Unauthorized'

    post:
      tags:
        - Campaigns
      summary: Create Campaign
      description: Create a new Campaign
      operationId: createCampaign
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateCampaignRequest'
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Campaign'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
...
  /api/v1/opportunities:
    get:
      tags:
        - Opportunities
      summary: List Opportunities
      description: Get a paginated list of Opportunities with optional filtering and sorting
      operationId: listOpportunities
      parameters:
        - $ref: '#/components/parameters/PageParam'
        - $ref: '#/components/parameters/PerPageParam'
        - $ref: '#/components/parameters/SortByParam'
        - $ref: '#/components/parameters/SortOrderParam'
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OpportunityList'
        '401':
          $ref: '#/components/responses/Unauthorized'

    post:
      tags:
        - Opportunities
      summary: Create Opportunity
      description: Create a new Opportunity
      operationId: createOpportunity
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateOpportunityRequest'
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Opportunity'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
...
components:
  schemas:
    Campaign:
      type: object
      required:
        - id
        - company_id
        - campaign_name
        - status
        - metadata
      properties:
        id:
          type: string
          format: uuid
        company_id:
          type: string
          format: uuid
        campaign_name:
          type: string
          maxLength: 140
        utm_source:
          nullable: true
          type: string
          maxLength: 80
        utm_medium:
          nullable: true
          type: string
          maxLength: 80
        utm_campaign:
          nullable: true
          type: string
          maxLength: 120
        status:
          type: string
        metadata:
          type: object
...
    CreateCampaignRequest:
      type: object
      required:
        - company_id
        - campaign_name
        - status
        - metadata
      properties:
        company_id:
          type: string
          format: uuid
        campaign_name:
          type: string
          maxLength: 140
        utm_source:
          nullable: true
          type: string
          maxLength: 80
        utm_medium:
          nullable: true
          type: string
          maxLength: 80
        utm_campaign:
          nullable: true
          type: string
          maxLength: 120
        status:
          type: string
        metadata:
          type: object
...
    CampaignList:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/Campaign'
        meta:
          $ref: '#/components/schemas/PaginationMeta'
    Opportunity:
      type: object
      required:
        - id
        - company_id
        - opportunity_name
        - currency
        - expected_amount
        - sales_stage
        - probability
        - status
        - metadata
      properties:
        id:
          type: string
          format: uuid
        company_id:
          type: string
          format: uuid
        opportunity_name:
          type: string
          maxLength: 140
        lead_id:
          nullable: true
          type: string
          format: uuid
        party_id:
          nullable: true
          type: string
          format: uuid
        campaign_id:
          nullable: true
          type: string
          format: uuid
        currency:
          type: string
        expected_amount:
          type: string
        sales_stage:
          type: string
        probability:
          type: string
        expected_close_date:
          nullable: true
          type: string
          format: date-time
        status:
          type: string
        quotation_id:
          nullable: true
          type: string
          format: uuid
        lost_reason:
          nullable: true
          type: string
          maxLength: 280
        competitor:
          nullable: true
          type: string
          maxLength: 140
        metadata:
          type: object
...
    CreateOpportunityRequest:
      type: object
      required:
        - company_id
        - opportunity_name
        - currency
        - expected_amount
        - sales_stage
        - probability
        - status
        - metadata
      properties:
        company_id:
          type: string
          format: uuid
        opportunity_name:
          type: string
          maxLength: 140
        lead_id:
          nullable: true
          type: string
          format: uuid
        party_id:
          nullable: true
          type: string
          format: uuid
        campaign_id:
          nullable: true
          type: string
          format: uuid
        currency:
          type: string
        expected_amount:
          type: string
        sales_stage:
          type: string
        probability:
          type: string
        expected_close_date:
          nullable: true
          type: string
          format: date-time
        status:
          type: string
        quotation_id:
          nullable: true
          type: string
          format: uuid
        lost_reason:
          nullable: true
          type: string
          maxLength: 280
        competitor:
          nullable: true
          type: string
          maxLength: 140
        metadata:
          type: object
...
    OpportunityList:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/Opportunity'
        meta:
          $ref: '#/components/schemas/PaginationMeta'
...
    PaginationMeta:
      type: object
      properties:
        total:
          type: integer
          example: 100
        page:
          type: integer
          example: 1
        per_page:
          type: integer
          example: 20
        total_pages:
          type: integer
          example: 5
    Error:
      type: object
      required:
        - code
        - message
      properties:
        code:
          type: string
          example: VALIDATION_ERROR
        message:
          type: string
          example: Invalid input data
        details:
          type: object
          additionalProperties: true
  parameters:
...
    PageParam:
      name: page
      in: query
      schema:
        type: integer
        minimum: 1
        default: 1
    PerPageParam:
      name: per_page
      in: query
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 20
    SortByParam:
      name: sort_by
      in: query
      schema:
        type: string
    SortOrderParam:
      name: sort_order
      in: query
      schema:
        type: string
        enum: [asc, desc]
        default: asc
  responses:
    BadRequest:
      description: Bad request
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    Unauthorized:
      description: Unauthorized
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
...
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
security:
  - bearerAuth: []