# src/exports/types.rs
//! Public types for Deal module
//!
//! Generated by metaphor-schema. Do not edit manually.
//!
//! These DTOs are the ONLY types other modules should use.
//! They are decoupled from internal domain entities.

use serde::{Deserialize, Serialize};
use uuid::Uuid;
use chrono::{DateTime, Utc};
use rust_decimal::Decimal;
use crate::domain::entity::*;

// ============================================================================
// CAMPAIGN TYPES
// ============================================================================

/// Type-safe ID for Campaign
///
/// Use this instead of raw Uuid for type safety across modules.
#[derive(Debug, Clone, Copy, PartialEq, Eq, Hash, Serialize, Deserialize)]
pub struct CampaignId(pub Uuid);

impl CampaignId {
    pub fn new(id: Uuid) -> Self {
        Self(id)
    }

    pub fn into_inner(self) -> Uuid {
        self.0
    }
}

impl From<Uuid> for CampaignId {
    fn from(id: Uuid) -> Self {
        Self(id)
    }
}

impl From<CampaignId> for Uuid {
    fn from(id: CampaignId) -> Self {
        id.0
    }
}

/// Data transfer object for Campaign
///
/// This is the public representation of Campaign for other modules.
#[derive(Debug, Clone, Serialize, Deserialize)]
pub struct CampaignDto {
    pub id: CampaignId,
    pub company_id: Uuid,
    pub campaign_name: String,
    pub utm_source: Option<String>,
    pub utm_medium: Option<String>,
    pub utm_campaign: Option<String>,
    pub status: CampaignStatus,
    pub metadata: serde_json::Value,
}

/// Summary view of Campaign for list displays
#[derive(Debug, Clone, Serialize, Deserialize)]
pub struct CampaignSummary {
    pub id: CampaignId,
    pub campaign_name: String,
    pub status: CampaignStatus,
}

/// Reference to Campaign for foreign key relationships
#[derive(Debug, Clone, Serialize, Deserialize)]
pub struct CampaignRef {
    pub id: CampaignId,
}

// ============================================================================
// OPPORTUNITY TYPES
// ============================================================================

/// Type-safe ID for Opportunity
///
/// Use this instead of raw Uuid for type safety across modules.
#[derive(Debug, Clone, Copy, PartialEq, Eq, Hash, Serialize, Deserialize)]
pub struct OpportunityId(pub Uuid);

impl OpportunityId {
    pub fn new(id: Uuid) -> Self {
        Self(id)
    }

    pub fn into_inner(self) -> Uuid {
        self.0
    }
}

impl From<Uuid> for OpportunityId {
    fn from(id: Uuid) -> Self {
        Self(id)
    }
}

impl From<OpportunityId> for Uuid {
    fn from(id: OpportunityId) -> Self {
        id.0
    }
}

/// Data transfer object for Opportunity
///
/// This is the public representation of Opportunity for other modules.
#[derive(Debug, Clone, Serialize, Deserialize)]
pub struct OpportunityDto {
    pub id: OpportunityId,
    pub company_id: Uuid,
    pub opportunity_name: String,
    pub lead_id: Option<Uuid>,
    pub party_id: Option<Uuid>,
    pub campaign_id: Option<Uuid>,
    pub currency: String,
    pub expected_amount: Decimal,
    pub sales_stage: SalesStage,
    pub probability: Decimal,
    pub expected_close_date: Option<DateTime<Utc>>,
    pub status: OpportunityStatus,
    pub quotation_id: Option<Uuid>,
    pub lost_reason: Option<String>,
    pub competitor: Option<String>,
    pub metadata: serde_json::Value,
}

/// Summary view of Opportunity for list displays
#[derive(Debug, Clone, Serialize, Deserialize)]
pub struct OpportunitySummary {
    pub id: OpportunityId,
    pub opportunity_name: String,
    pub status: OpportunityStatus,
}

/// Reference to Opportunity for foreign key relationships
#[derive(Debug, Clone, Serialize, Deserialize)]
pub struct OpportunityRef {
    pub id: OpportunityId,
}

// ============================================================================
// OPPORTUNITYITEM TYPES
// ============================================================================

/// Type-safe ID for OpportunityItem
///
/// Use this instead of raw Uuid for type safety across modules.
#[derive(Debug, Clone, Copy, PartialEq, Eq, Hash, Serialize, Deserialize)]
pub struct OpportunityItemId(pub Uuid);

impl OpportunityItemId {
    pub fn new(id: Uuid) -> Self {
        Self(id)
    }

    pub fn into_inner(self) -> Uuid {
        self.0
    }
}

impl From<Uuid> for OpportunityItemId {
    fn from(id: Uuid) -> Self {
        Self(id)
    }
}

impl From<OpportunityItemId> for Uuid {
    fn from(id: OpportunityItemId) -> Self {
        id.0
    }
}

/// Data transfer object for OpportunityItem
///
/// This is the public representation of OpportunityItem for other modules.
#[derive(Debug, Clone, Serialize, Deserialize)]
pub struct OpportunityItemDto {
    pub id: OpportunityItemId,
    pub opportunity_id: Uuid,
    pub company_id: Uuid,
    pub item_id: Uuid,
    pub description: Option<String>,
    pub quantity: Decimal,
    pub rate: Decimal,
    pub amount: Decimal,
    pub metadata: serde_json::Value,
}

/// Summary view of OpportunityItem for list displays
#[derive(Debug, Clone, Serialize, Deserialize)]
pub struct OpportunityItemSummary {
    pub id: OpportunityItemId,
}

/// Reference to OpportunityItem for foreign key relationships
#[derive(Debug, Clone, Serialize, Deserialize)]
pub struct OpportunityItemRef {
    pub id: OpportunityItemId,
}

// ============================================================================
// CUSTOM TYPES
// ============================================================================

// <<< CUSTOM TYPES START >>>
// Add custom public types here
// <<< CUSTOM TYPES END >>>